+++ apps/dokploy/pages/dashboard/project/[projectId]/environment/[environmentId]/services/application/[applicationId].tsx
@@ -1,3 +1,4 @@
+import { ENABLE_TRAEFIK } from "@dokploy/server/constants/env";
 import { validateRequest } from "@dokploy/server/lib/auth";
 import { createServerSideHelpers } from "@trpc/react-query/server";
 import copy from "copy-to-clipboard";
@@ -239,7 +240,7 @@ const Service = (
 													Environment
 												</TabsTrigger>
 											)}
-											{permissions?.domain.read && (
+											{permissions?.domain.read && ENABLE_TRAEFIK && (
 												<TabsTrigger value="domains">Domains</TabsTrigger>
 											)}
 											{permissions?.deployment.read && (
@@ -395,7 +396,7 @@ const Service = (
 											</div>
 										</TabsContent>
 									)}
-									{permissions?.domain.read && (
+									{permissions?.domain.read && ENABLE_TRAEFIK && (
 										<TabsContent value="domains" className="w-full">
 											<div className="flex flex-col gap-4 pt-2.5">
 												<ShowDomains id={applicationId} type="application" />
@@ -418,10 +419,10 @@ const Service = (
 												<ShowBuildServer applicationId={applicationId} />
 												<ShowResources id={applicationId} type="application" />
 												<ShowVolumes id={applicationId} type="application" />
-												<ShowRedirects applicationId={applicationId} />
-												<ShowSecurity applicationId={applicationId} />
+												{ENABLE_TRAEFIK && <ShowRedirects applicationId={applicationId} />}
+												{ENABLE_TRAEFIK && <ShowSecurity applicationId={applicationId} />}
 												<ShowPorts applicationId={applicationId} />
-												<ShowTraefikConfig applicationId={applicationId} />
+												{ENABLE_TRAEFIK && <ShowTraefikConfig applicationId={applicationId} />}
 											</div>
 										</TabsContent>
 									)}